requirements-dev.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
tiktoken
